chore: split to @components, @hooks, etc.
