+++ dropbox-hacking-photo-manager-client/src/components/editableTextField.tsx
@@ -9,7 +9,7 @@ import {
   useState,
 } from "react";
 
-import logRender from "@/logRender";
+import logRender from "@lib/logRender";
 
 const EditableTextField = (props: {
   value: string;
