+++ dropbox-hacking-photo-manager-client/src/hooks/useThumbnail/types.ts
