+++ dropbox-hacking-photo-manager-client/src/hooks/useThumbnail/websocketThumbnailLoader.ts
