+++ dropbox-hacking-photo-manager-client/src/index.tsx
@@ -9,26 +9,26 @@ import { createRoot } from "react-dom/client";
 
 import * as multiplexerContext from "./context/rx/multiplexerContext";
 
-import { context as routingContext, type Router } from "@/hooks/useRouter";
+import { context as routingContext, type Router } from "@hooks/useRouter";
 import * as rxRecordFeedContext from "./context/rx/rxRecordFeedContext";
 import * as additionalFeeds from "./context/rx/additionalFeeds";
-import * as thumbnailLoaderContext from "./context/thumbnails";
-import Day from "@/pages/legacy/day";
-import Calendar from "@/pages/legacy/days/calendar";
-import ListOfDays from "@/pages/legacy/days/listOfDays";
-import logRender from "@/logRender";
-import Photo from "@/pages/legacy/photo";
-import ClosestTo from "@/pages/legacy/closest-to/index";
-import Month from "@/pages/legacy/month";
-import Year from "@/pages/legacy/year";
-import BasicCounts from "@/pages/next-gen/basic-counts";
-import NGDaysNoSamples from "@/pages/next-gen/list-of-days/without-samples";
-import NGDayFiles from "@/pages/next-gen/day/files";
-import NGContentHash from "@/pages/next-gen/contentHash";
-import NGFileId from "@/pages/next-gen/fileId";
-import NGFileRev from "@/pages/next-gen/fileRev";
-import Fsck from "@/pages/next-gen/fsck";
-import ExifExplorer from "@/pages/next-gen/exifExplorer";
+import { defaultProvider as ThumbnailProvider } from "@hooks/useThumbnail";
+import Day from "@pages/legacy/day";
+import Calendar from "@pages/legacy/days/calendar";
+import ListOfDays from "@pages/legacy/days/listOfDays";
+import logRender from "@lib/logRender";
+import Photo from "@pages/legacy/photo";
+import ClosestTo from "@pages/legacy/closest-to/index";
+import Month from "@pages/legacy/month";
+import Year from "@pages/legacy/year";
+import BasicCounts from "@pages/next-gen/basic-counts";
+import NGDaysNoSamples from "@pages/next-gen/list-of-days/without-samples";
+import NGDayFiles from "@pages/next-gen/day/files";
+import NGContentHash from "@pages/next-gen/contentHash";
+import NGFileId from "@pages/next-gen/fileId";
+import NGFileRev from "@pages/next-gen/fileRev";
+import Fsck from "@pages/next-gen/fsck";
+import ExifExplorer from "@pages/next-gen/exifExplorer";
 
 const ensureNever = <_ extends never>() => undefined;
 
@@ -141,9 +141,7 @@ const Root = ({
   return (
     <routingContext.Provider value={router}>
       <multiplexerContext.defaultProvider accepter={accepter}>
-        <thumbnailLoaderContext.defaultProvider>
-          {toRender({ routeState: state })}
-        </thumbnailLoaderContext.defaultProvider>
+        <ThumbnailProvider>{toRender({ routeState: state })}</ThumbnailProvider>
       </multiplexerContext.defaultProvider>
     </routingContext.Provider>
   );
