+++ dropbox-hacking-photo-manager-client/src/pages/legacy/day/index.tsx
@@ -6,14 +6,14 @@ import * as React from "react";
 import { useEffect, useMemo, useRef, useState } from "react";
 
 import useVisibilityTracking from "../days/useVisibilityTracking";
-import logRender from "@/logRender";
-import EditableTextField from "@/components/editableTextField";
-import PhotoTile from "@/components/photoTile";
+import logRender from "@lib/logRender";
+import EditableTextField from "@components/editableTextField";
+import PhotoTile from "@components/photoTile";
 import { useRxFeedsViaMultiplexer } from "@/context/rx/rxRecordFeedContext";
 import { map } from "rxjs";
 import { useLatestValue } from "@/context/rx/useLatestValue";
 import { useAdditionalFeeds } from "@/context/rx/additionalFeeds";
-import SamePageLink from "@/components/samePageLink";
+import SamePageLink from "@components/samePageLink";
 
 const DayWithData = ({
   date,
