+++ dropbox-hacking-photo-manager-client/src/pages/legacy/days/calendar/index.tsx
@@ -2,9 +2,9 @@ import * as React from "react";
 import { useEffect } from "react";
 
 import { useCountsByDate } from "@/context/countsByDateContext";
-import logRender from "@/logRender";
-import SamePageLink from "@/components/samePageLink";
-import Navigate from "@/components/navigate";
+import logRender from "@lib/logRender";
+import SamePageLink from "@components/samePageLink";
+import Navigate from "@components/navigate";
 import MonthBox from "./monthBox";
 import TwelveMonths from "./twelveMonths";
