+++ dropbox-hacking-photo-manager-client/src/pages/legacy/days/calendar/twelveMonths.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 
-import logRender from "@/logRender";
+import logRender from "@lib/logRender";
 
 const TwelveMonths = (props: {
   year: number;
