+++ dropbox-hacking-photo-manager-client/src/pages/legacy/days/dayWithSamples.tsx
@@ -2,9 +2,9 @@ import { CountsByDateEntry } from "dropbox-hacking-photo-manager-shared";
 import * as React from "react";
 import { useEffect, useMemo, useState } from "react";
 
-import EditableTextField from "@/components/editableTextField";
-import logRender from "@/logRender";
-import SamePageLink from "@/components/samePageLink";
+import EditableTextField from "@components/editableTextField";
+import logRender from "@lib/logRender";
+import SamePageLink from "@components/samePageLink";
 import type { Subscribable } from "./emittableSubscribable";
 import SamplePhoto from "./samplePhoto";
