+++ dropbox-hacking-photo-manager-client/src/pages/legacy/days/listOfDays.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import { useEffect } from "react";
 
 import { useCountsByDate } from "@/context/countsByDateContext";
-import logRender from "@/logRender";
+import logRender from "@lib/logRender";
 import ListOfDaysWithData from "./listOfDaysWithData";
 import { useLatestValue } from "@/context/rx/useLatestValue";
 import { useRxFeedsViaMultiplexer } from "@/context/rx/rxRecordFeedContext";
