+++ dropbox-hacking-photo-manager-client/src/pages/legacy/days/samplePhoto.tsx
@@ -1,9 +1,8 @@
 import { Photo } from "dropbox-hacking-photo-manager-shared";
 import * as React from "react";
 
-import { useThumbnailLoader } from "@/context/thumbnails";
-import { useThumbnail } from "@/context/thumbnails/useThumbnail";
-import logRender from "@/logRender";
+import useThumbnail from "@hooks/useThumbnail";
+import logRender from "@lib/logRender";
 
 const samplePhoto = ({
   photo,
@@ -12,8 +11,7 @@ const samplePhoto = ({
   photo: Photo;
   visible: boolean;
 }): React.ReactElement | null => {
-  const loader = useThumbnailLoader();
-  const thumbnail = useThumbnail(photo.namedFile.rev, loader)(visible);
+  const thumbnail = useThumbnail(photo.namedFile.rev)(visible);
 
   return (
     <span key={photo.namedFile.id} className="sample">
