+++ dropbox-hacking-photo-manager-client/src/pages/legacy/photo/index.tsx
@@ -2,11 +2,11 @@ import { GPSLatLong, Photo } from "dropbox-hacking-photo-manager-shared";
 import * as React from "react";
 import { useEffect, useMemo, useState } from "react";
 
-import logRender from "@/logRender";
+import logRender from "@lib/logRender";
 import { useLatestValue } from "@/context/rx/useLatestValue";
 import { useAdditionalFeeds } from "@/context/rx/additionalFeeds";
-import SamePageLink from "@/components/samePageLink";
-import EditableTextField from "@/components/editableTextField";
+import SamePageLink from "@components/samePageLink";
+import EditableTextField from "@components/editableTextField";
 
 const Photo = (props: { rev: string }): React.ReactElement | null => {
   const [previewSizes, setPreviewSizes] = useState<string[]>();
