+++ dropbox-hacking-photo-manager-client/src/pages/legacy/stats/index.tsx
@@ -2,10 +2,10 @@ import * as React from "react";
 import { useEffect, useMemo, useState } from "react";
 import { map, Observable } from "rxjs";
 
-import logRender from "@/logRender";
+import logRender from "@lib/logRender";
 import { useRxFeedsViaMultiplexer } from "@/context/rx/rxRecordFeedContext";
 import type { ImageAndMaybeDelta } from "@/context/rx/rxFeedClient";
-import Navigate from "@/components/navigate";
+import Navigate from "@components/navigate";
 import { useLatestValue } from "@/context/rx/useLatestValue";
 import { useAdditionalFeeds } from "@/context/rx/additionalFeeds";
