+++ dropbox-hacking-photo-manager-client/src/pages/next-gen/contentHash/index.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from "react";
 
-import logRender from "@/logRender";
-import Navigate from "@/components/navigate";
+import logRender from "@lib/logRender";
+import Navigate from "@components/navigate";
 import type { ContentHashResult } from "dropbox-hacking-photo-manager-shared/serverSideFeeds";
 import { useLatestValueFromServerFeed } from "../useLatestValueFromServerFeed";
 import SummariseMediaInfo from "./SummariseMediaInfo";
