+++ dropbox-hacking-photo-manager-client/src/pages/next-gen/day/filesTable.tsx
@@ -1,9 +1,9 @@
 import React, { useEffect, useRef, useState } from "react";
 
-import logRender from "@/logRender";
+import logRender from "@lib/logRender";
 import type { DayFilesResult } from "dropbox-hacking-photo-manager-shared/serverSideFeeds";
-import SamePageLink from "@/components/samePageLink";
-import useVisibilityTracking from "@/pages/legacy/days/useVisibilityTracking";
+import SamePageLink from "@components/samePageLink";
+import useVisibilityTracking from "@pages/legacy/days/useVisibilityTracking";
 import MaybeVisibleThumbnail from "./MaybeVisibleThumbnail";
 import { Observable } from "rxjs";
