+++ dropbox-hacking-photo-manager-client/src/pages/next-gen/fileRev/index.tsx
@@ -1,10 +1,10 @@
 import React, { useEffect } from "react";
 
-import logRender from "@/logRender";
-import Navigate from "@/components/navigate";
+import logRender from "@lib/logRender";
+import Navigate from "@components/navigate";
 import type { FileRevResult } from "dropbox-hacking-photo-manager-shared/serverSideFeeds";
 import { useLatestValueFromServerFeed } from "../useLatestValueFromServerFeed";
-import SamePageLink from "@/components/samePageLink";
+import SamePageLink from "@components/samePageLink";
 
 const NGFileRev = ({ rev }: { rev: string }) => {
   const latestValue = useLatestValueFromServerFeed<FileRevResult>({
