+++ dropbox-hacking-photo-manager-client/src/pages/next-gen/list-of-days/without-samples.tsx
@@ -1,10 +1,10 @@
 import React, { useEffect } from "react";
 
-import logRender from "@/logRender";
-import Navigate from "@/components/navigate";
+import logRender from "@lib/logRender";
+import Navigate from "@components/navigate";
 import type { NGDaysNoSamplesType } from "dropbox-hacking-photo-manager-shared/serverSideFeeds";
 import { useLatestValueFromServerFeed } from "../useLatestValueFromServerFeed";
-import SamePageLink from "@/components/samePageLink";
+import SamePageLink from "@components/samePageLink";
 
 const NGDaysNoSamples = () => {
   const latestValue = useLatestValueFromServerFeed<NGDaysNoSamplesType>({
