+++ dropbox-hacking-photo-manager-client/tsconfig.json
@@ -9,7 +9,11 @@
     "noImplicitThis": true,
     "outDir": "./public/dist/",
     "paths": {
-      "@/*": ["./*"]
+      "@/*": ["./*"],
+      "@components/*": ["./components/*"],
+      "@hooks/*": ["./hooks/*"],
+      "@lib/*": ["./lib/*"],
+      "@pages/*": ["./pages/*"]
     },
     "resolveJsonModule": true,
     "sourceMap": true,
