+++ dropbox-hacking-photo-manager-client/webpack.config.ts
@@ -43,7 +43,11 @@ const config: webpack.Configuration = {
   ],
   resolve: {
     alias: {
-      "@/*": path.resolve(import.meta.dirname, "src/*"),
+      "@/*": path.resolve(__dirname, "src/*"),
+      "@components/*": path.resolve(__dirname, "src/components/*"),
+      "@hooks/*": path.resolve(__dirname, "src/hooks/*"),
+      "@lib/*": path.resolve(__dirname, "src/lib/*"),
+      "@pages/*": path.resolve(__dirname, "src/pages/*"),
     },
     extensions: [".js", ".ts", ".tsx"],
   },
